refactor(QuestionTest, QuestionRepositoryTest): 통합 테스트(DB)와 도메인 단위 테스트 분리
